requirements.txt
boto3==1.13.18
coloredlogs
dlib
google-cloud-storage
gunicorn==20.0.4
h5py
numpy==1.18.1
opencv-python==4.2.0.34
pandas==1.0.3
pre-commit==1.14.2
python-dotenv==0.13.0
scikit-learn==0.22.1
tensorflow==1.15.0
tornado==5.1.1
tqdm==4.44.1
ujson
watchtower==0.7.3
xgboost==1.0.2
scipy==1.4.1
joblib==0.14.1
memory-profiler